# docs/swagger.yaml
basePath: /
definitions:
  dto.ChangePasswordRequest:
    properties:
      '-':
        type: string
      current:
        type: string
      new:
        type: string
      revoke_other_session:
        type: boolean
    required:
    - current
    - new
    - revoke_other_session
    type: object
  dto.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  dto.GetMeResponse:
    properties:
      user:
        $ref: '#/definitions/dto.UserProfileData'
    type: object
  dto.LoginRequest:
    properties:
      identifier:
        type: string
      password:
        type: string
    required:
    - identifier
    - password
    type: object
  dto.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/dto.UserData'
    type: object
  dto.RefreshTokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      token_type:
        type: string
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
      username:
        type: string
    required:
    - email
    - name
    - password
    - username
    type: object
  dto.RegisterResponse:
    properties:
      user:
        $ref: '#/definitions/dto.UserData'
    type: object
  dto.ResetPasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
  dto.UpdateProfileData:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      name:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  dto.UpdateProfileResponse:
    properties:
      user:
        $ref: '#/definitions/dto.UpdateProfileData'
    type: object
  dto.UserData:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      name:
        type: string
      status:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  dto.UserDeleteAccountRequest:
    properties:
      '-':
        type: string
      password:
        type: string
    type: object
  dto.UserProfileData:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      email_verified_at:
        type: string
      last_login_at:
        type: string
      name:
        type: string
      status:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  lib.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        default: true
        type: boolean
    type: object
  lib.ErrorResponse:
    properties:
      code:
        type: string
      details: {}
    type: object
  lib.HTTPError:
    properties:
      error:
        $ref: '#/definitions/lib.ErrorResponse'
      message:
        type: string
      success:
        default: false
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Api starter template for lazy people
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go gin starter template API
  version: "1.0"
paths:
  /api/v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send reset password link to email
      parameters:
      - description: Forgot Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lib.APIResponse'
            - properties:
                message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.HTTPError'
      summary: Forgot password
      tags:
      - Auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Login user and get tokens
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lib.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.HTTPError'
      summary: Login user
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke current user session and invalidate the refresh token in
        database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.HTTPError'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - Auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using a valid refresh token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lib.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RefreshTokenResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.HTTPError'
      security:
      - BearerAuth: []
      summary: Refresh access token
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: Register Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/lib.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RegisterResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.HTTPError'
      summary: Register a new user
      tags:
      - Auth
  /api/v1/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset user password using a reset token
      parameters:
      - description: Reset Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.HTTPError'
      summary: Reset user password
      tags:
      - Auth
  /api/v1/users/me:
    delete:
      consumes:
      - application/json
      description: Deletes the authenticated user's account.
      parameters:
      - description: Delete account request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserDeleteAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Account deleted successfully
          schema:
            $ref: '#/definitions/lib.APIResponse'
        "400":
          description: Invalid password or ownership transfer required
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/lib.HTTPError'
      security:
      - BearerAuth: []
      summary: Delete user account
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Fetches the detailed profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved profile
          schema:
            allOf:
            - $ref: '#/definitions/lib.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetMeResponse'
              type: object
        "401":
          description: 'Unauthorized: Missing or invalid token'
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "403":
          description: 'Forbidden: User is not associated with the requested organization'
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "404":
          description: 'Not Found: User record does not exist'
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "500":
          description: 'Internal Server Error: Something went wrong on our end'
          schema:
            $ref: '#/definitions/lib.HTTPError'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - Users
    put:
      consumes:
      - multipart/form-data
      description: Updates the name and avatar URL of the authenticated user.
      parameters:
      - description: New name of the user
        in: formData
        name: name
        type: string
      - description: User avatar image file
        in: formData
        name: avatar
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/lib.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UpdateProfileResponse'
              type: object
        "400":
          description: 'Bad Request: Invalid input'
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.HTTPError'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Users
  /api/v1/users/me/password:
    put:
      consumes:
      - application/json
      description: Allows an authenticated user to change their password and optionally
        revoke other active sessions.
      parameters:
      - description: Change Password Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
          schema:
            $ref: '#/definitions/lib.APIResponse'
        "400":
          description: 'Bad Request: Validation error or password mismatch'
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "401":
          description: 'Unauthorized: Missing or invalid token'
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.HTTPError'
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"